+++ packages/ph-cli/src/commands/index.ts
@@ -1,11 +1,13 @@
 import { Command } from "commander";
+import { connectCommand } from "./connect.js";
 import { devCommand } from "./dev.js";
+import { generateCommand } from "./generate.js";
 import { helpCommand } from "./help.js";
 import { initCommand } from "./init.js";
-import { connectCommand } from "./connect.js";
+import { installCommand } from "./install.js";
 import { reactorCommand } from "./reactor.js";
-import { generateCommand } from "./generate.js";
 
+import { serviceCommand } from "./service.js";
 export const commands = [
   initCommand,
   devCommand,
@@ -13,6 +15,8 @@ export const commands = [
   generateCommand,
   reactorCommand,
   helpCommand,
+  installCommand,
+  serviceCommand,
 ];
 
 export default function registerCommands(program: Command) {
@@ -20,7 +24,8 @@ export default function registerCommands(program: Command) {
 }
 
 export * from "./dev.js";
+export * from "./generate.js";
 export * from "./help.js";
 export * from "./init.js";
-export * from "./generate.js";
 export * from "./reactor.js";
+export * from "./service.js";