+++ packages/ph-cmd/src/commands/index.ts
@@ -1,0 +1,10 @@
+import { Command } from "commander";
+import { setupGlobalsCommand } from "./setup-globals.js";
+
+export const commands = [setupGlobalsCommand];
+
+export default function registerCommands(program: Command) {
+  commands.forEach((command) => command(program));
+}
+
+export * from "./setup-globals.js";